pyproject.toml
[project]
name = "todo-teste"
version = "0.1.0"
description = ""
authors = [
    {name = "Danielle Rocha",email = "daniellerocha.c@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13,<4.0"
dependencies = [
    "fastapi[standard] (>=0.115.12,<0.116.0)",
    "python-multipart (>=0.0.20,<0.0.21)",
    "pyjwt (>=2.10.1,<3.0.0)",
    "sqlalchemy (>=2.0.41,<3.0.0)",
    "pydantic-settings (>=2.9.1,<3.0.0)",
    "alembic (>=1.16.1,<2.0.0)",
    "pwdlib[argon2] (>=0.2.1,<0.3.0)",
    "tzdata (>=2025.2,<2026.0)"
]

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.5"
pytest-cov = "^6.1.1"
taskipy = "^1.14.1"
ruff = "^0.11.10"

[tool.ruff]
line-length = 88
extend-exclude = ['migrations']


[tool.ruff.lint]
preview = true
select = ['I', 'F', 'E', 'W', 'PL', 'PT']


[tool.ruff.format]
preview = true
quote-style = 'single'

[tool.pytest.ini_options]
pythonpath = "."
addopts = '-p no:warnings'


[tool.taskipy.tasks]
run = 'fastapi dev todo_teste/app.py'


pre_test = 'task lint'
test = 'pytest --cov=todo_teste -vv'
post_test = 'coverage html'


lint = 'ruff check'
pre_format = 'ruff check --fix'
format = 'ruff format'

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"


